requirements.txt
aiogram==2.21
nest_asyncio==1.5.5
python-decouple==3.6
xlsxwritter==3.0.8
SQLAlchemy==2.0.1
